클라이언트 배포 api 서버 반영

+++ server/clothes/views.py
@@ -1,5 +1,3 @@
-import os
-import sys
 from drf_yasg import openapi
 from drf_yasg.utils import swagger_auto_schema
 from rest_framework.views import APIView
@@ -9,10 +7,7 @@ from .models import Clothes
 from .serializers import ClothesSerializer, ClorareSerializer
 from . import swagger as sw
 from .utils import dec_b64_img, modifying_image_path
-
-p = os.path.abspath('.') + r'\openweather'
-sys.path.insert(1, p)
-from openweatherapi import open_weather_api
+from openweather.openweatherapi import open_weather_api
 
 
 # Create your views here.
@@ -30,14 +25,26 @@ class ClothesView(APIView):
                     'gender': sw.cms_gender,
                     'image': sw.cms_image
                 })
+            ),
+            status.HTTP_400_BAD_REQUEST: openapi.Response(
+                'Failed', schema=openapi.Schema(type=openapi.TYPE_OBJECT, properties={
+                    'message': sw.sms_400
+                })
+            ),
+            status.HTTP_500_INTERNAL_SERVER_ERROR: openapi.Response(
+                'Failed', schema=openapi.Schema(type=openapi.TYPE_OBJECT, properties={
+                    'message': sw.sms_500
+                })
             )
         }
     )
     def get(self, request):
         try:
-            queryset = Clothes.objects.all()
-            serializer = ClothesSerializer(queryset, many=True)
-            return Response(serializer.data, status=status.HTTP_200_OK)
+            if request.user.is_authenticated:
+                queryset = Clothes.objects.all()
+                serializer = ClothesSerializer(queryset, many=True)
+                return Response(serializer.data, status=status.HTTP_200_OK)
+            return Response({'message': "You are not a logged in user"}, status=status.HTTP_400_BAD_REQUEST)
         except Exception as e:
             return Response({'message': str(e)})
 
@@ -70,23 +77,30 @@ class ClothesView(APIView):
                 'Failed', schema=openapi.Schema(type=openapi.TYPE_OBJECT, properties={
                     'message': sw.sms_400
                 })
+            ),
+            status.HTTP_500_INTERNAL_SERVER_ERROR: openapi.Response(
+                'Failed', schema=openapi.Schema(type=openapi.TYPE_OBJECT, properties={
+                    'message': sw.sms_500
+                })
             )
         }
     )
     def post(self, request):
         try:
-            context = {
-                'category': request.data['category'],
-                'cloth_type': request.data['cloth_type'],
-                'season': request.data['season'],
-                'gender': request.data['gender'],
-                'image': dec_b64_img(request.data)
-            }
-            serializer = ClothesSerializer(data=context)
-            if serializer.is_valid():
-                queryset = serializer.save()
-                return Response(serializer.data, status=status.HTTP_201_CREATED)
-            return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
+            if request.user.is_authenticated:
+                context = {
+                    'category': request.data['category'],
+                    'cloth_type': request.data['cloth_type'],
+                    'season': request.data['season'],
+                    'gender': request.data['gender'],
+                    'image': dec_b64_img(request.data)
+                }
+                serializer = ClothesSerializer(data=context)
+                if serializer.is_valid():
+                    queryset = serializer.save()
+                    return Response(serializer.data, status=status.HTTP_201_CREATED)
+                return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
+            return Response({'message': "You are not a logged in user"}, status=status.HTTP_400_BAD_REQUEST)
         except Exception as e:
             return Response({'message': str(e)})
 
@@ -107,6 +121,11 @@ class ClothesDetailView(APIView):
                     'image': sw.cms_image
                 })
             ),
+            status.HTTP_400_BAD_REQUEST: openapi.Response(
+                'Failed', schema=openapi.Schema(type=openapi.TYPE_OBJECT, properties={
+                    'message': sw.sms_400
+                })
+            ),
             status.HTTP_500_INTERNAL_SERVER_ERROR: openapi.Response(
                 'Failed', schema=openapi.Schema(type=openapi.TYPE_OBJECT, properties={
                     'message': sw.sms_500
@@ -116,9 +135,11 @@ class ClothesDetailView(APIView):
     )
     def get(self, request, pk):
         try:
-            queryset = Clothes.objects.get(pk=pk)
-            serializer = ClothesSerializer(queryset)
-            return Response(serializer.data, status=status.HTTP_200_OK)
+            if request.user.is_authenticated:
+                queryset = Clothes.objects.get(pk=pk)
+                serializer = ClothesSerializer(queryset)
+                return Response(serializer.data, status=status.HTTP_200_OK)
+            return Response({'message': "You are not a logged in user"}, status=status.HTTP_400_BAD_REQUEST)
         except Exception as e:
             return Response({'message': str(e)})
 
@@ -162,12 +183,14 @@ class ClothesDetailView(APIView):
     )
     def put(self, request, pk):
         try:
-            queryset = Clothes.objects.get(pk=pk)
-            serializer = ClothesSerializer(queryset, data=request.data)
-            if serializer.is_valid():
-                serializer.save()
-                return Response(serializer.data, status=status.HTTP_202_ACCEPTED)
-            return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
+            if request.user.is_authenticated:
+                queryset = Clothes.objects.get(pk=pk)
+                serializer = ClothesSerializer(queryset, data=request.data)
+                if serializer.is_valid():
+                    serializer.save()
+                    return Response(serializer.data, status=status.HTTP_202_ACCEPTED)
+                return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
+            return Response({'message': "You are not a logged in user"}, status=status.HTTP_400_BAD_REQUEST)
         except Exception as e:
             return Response({'message': str(e)})
 
@@ -181,6 +204,11 @@ class ClothesDetailView(APIView):
                     'message': sw.sms_202
                 })
             ),
+            status.HTTP_400_BAD_REQUEST: openapi.Response(
+                'Failed', schema=openapi.Schema(type=openapi.TYPE_OBJECT, properties={
+                    'message': sw.sms_400
+                })
+            ),
             status.HTTP_500_INTERNAL_SERVER_ERROR: openapi.Response(
                 'Failed', schema=openapi.Schema(type=openapi.TYPE_OBJECT, properties={
                     'message': sw.sms_500
@@ -190,9 +218,11 @@ class ClothesDetailView(APIView):
     )
     def delete(self, request, pk):
         try:
-            queryset = Clothes.objects.get(pk=pk)
-            queryset.delete()
-            return Response({'message': 'Cloth deleted'}, status=status.HTTP_202_ACCEPTED)
+            if request.user.is_authenticated:
+                queryset = Clothes.objects.get(pk=pk)
+                queryset.delete()
+                return Response({'message': 'Cloth deleted'}, status=status.HTTP_202_ACCEPTED)
+            return Response({'message': "You are not a logged in user"}, status=status.HTTP_400_BAD_REQUEST)
         except Exception as e:
             return Response({'message': str(e)})
 
@@ -201,7 +231,6 @@ class ClothesRecommendationView(APIView):
     @swagger_auto_schema(
         operation_id='Cloth recommendation',
         operation_description='Recommend clothes at random',
-        manual_parameters=[sw.rp_gender],
         responses={
             status.HTTP_200_OK: openapi.Response(
                 'Success', schema=openapi.Schema(type=openapi.TYPE_OBJECT, properties={
@@ -237,44 +266,56 @@ class ClothesRecommendationView(APIView):
                         'wind_speed': sw.wms_wind_speed
                     })
                 })
+            ),
+            status.HTTP_400_BAD_REQUEST: openapi.Response(
+                'Failed', schema=openapi.Schema(type=openapi.TYPE_OBJECT, properties={
+                    'message': sw.sms_400
+                })
+            ),
+            status.HTTP_500_INTERNAL_SERVER_ERROR: openapi.Response(
+                'Failed', schema=openapi.Schema(type=openapi.TYPE_OBJECT, properties={
+                    'message': sw.sms_500
+                })
             )
         }
     )
     def get(self, request, gender):
         try:
-            gender_str = ["female", "male"]
-            weather_data = open_weather_api()[1]
+            if request.user.is_authenticated:
+                gender = request.user.gender
+                weather_data = open_weather_api()[1]
 
-            if weather_data["temperature"] >= 30:
-                season = "summer"
-            elif weather_data["temperature"] >= 20:
-                season = "autumn"
-            elif weather_data["temperature"] >= 10:
-                season = "spring"
-            else:
-                season = "winter"
+                if weather_data["temperature"] >= 30:
+                    season = "summer"
+                elif weather_data["temperature"] >= 20:
+                    season = "autumn"
+                elif weather_data["temperature"] >= 10:
+                    season = "spring"
+                else:
+                    season = "winter"
 
-            queryset = Clothes.objects.exclude(gender=gender_str[1-int(gender)])
-            queryset = queryset.filter(season=season)
+                queryset = Clothes.objects.filter(gender=gender) | Clothes.objects.filter(gender="unisex")
+                queryset = queryset.filter(season=season)
 
-            cloth_top = queryset.filter(category="top").order_by("?").first()
-            cloth_bottom = queryset.filter(category="bottom").order_by("?").first()
-            cloth_coat = queryset.filter(category="coat").order_by("?").first()
+                cloth_top = queryset.filter(category="top").order_by("?").first()
+                cloth_bottom = queryset.filter(category="bottom").order_by("?").first()
+                cloth_coat = queryset.filter(category="coat").order_by("?").first()
 
-            serialized_top = ClothesSerializer(cloth_top).data
-            serialized_bottom = ClothesSerializer(cloth_bottom).data
-            serialized_coat = ClothesSerializer(cloth_coat).data
+                serialized_top = ClothesSerializer(cloth_top).data
+                serialized_bottom = ClothesSerializer(cloth_bottom).data
+                serialized_coat = ClothesSerializer(cloth_coat).data
 
-            context = {
-                "top": modifying_image_path(serialized_top),
-                "bot": modifying_image_path(serialized_bottom),
-                "coat": modifying_image_path(serialized_coat),
-                "gender": gender_str[gender],
-                "w_data": weather_data
-            }
+                context = {
+                    "top": modifying_image_path(serialized_top),
+                    "bot": modifying_image_path(serialized_bottom),
+                    "coat": modifying_image_path(serialized_coat),
+                    "gender": gender,
+                    "w_data": weather_data
+                }
 
-            serializer = ClorareSerializer(context)
-            return Response(serializer.data, status=status.HTTP_200_OK)
+                serializer = ClorareSerializer(context)
+                return Response(serializer.data, status=status.HTTP_200_OK)
+            return Response({'message': "You are not a logged in user"}, status=status.HTTP_400_BAD_REQUEST)
         except Exception as e:
             return Response({'message': str(e)})
 
@@ -296,20 +337,32 @@ class ClothesCleanUpView(APIView):
                     'message': sw.sms_202
                 })
             ),
+            status.HTTP_400_BAD_REQUEST: openapi.Response(
+                'Failed', schema=openapi.Schema(type=openapi.TYPE_OBJECT, properties={
+                    'message': sw.sms_400
+                })
+            ),
             status.HTTP_403_FORBIDDEN: openapi.Response(
                 'Failed', schema=openapi.Schema(type=openapi.TYPE_OBJECT, properties={
                     'message': sw.sms_403
+                })
+            ),
+            status.HTTP_500_INTERNAL_SERVER_ERROR: openapi.Response(
+                'Failed', schema=openapi.Schema(type=openapi.TYPE_OBJECT, properties={
+                    'message': sw.sms_500
                 })
             )
         }
     )
     def post(self, request):
         try:
-            if request.data['key'] == "delete all clothes data":
-                queryset = Clothes.objects.all()
-                queryset.delete()
-                return Response({'message': 'Cloth deleted'}, status=status.HTTP_202_ACCEPTED)
-            return Response({'message': 'Incorrect delete key'}, status=status.HTTP_403_FORBIDDEN)
+            if request.user.is_authenticated:
+                if request.data['key'] == "delete all clothes data":
+                    queryset = Clothes.objects.all()
+                    queryset.delete()
+                    return Response({'message': 'Cloth deleted'}, status=status.HTTP_202_ACCEPTED)
+                return Response({'message': 'Incorrect delete key'}, status=status.HTTP_403_FORBIDDEN)
+            return Response({'message': "You are not a logged in user"}, status=status.HTTP_400_BAD_REQUEST)
         except Exception as e:
             return Response({'message': str(e)})
